#1798 decide column count only by the first line
