+++ discovery-server/src/main/java/app/metatron/discovery/domain/dataprep/PrepUtil.java
@@ -16,8 +16,9 @@ package app.metatron.discovery.domain.dataprep;
 
 import static app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes.PREP_DATAFLOW_ERROR_CODE;
 import static app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes.PREP_DATASET_ERROR_CODE;
+import static app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes.PREP_INVALID_CONFIG_CODE;
+import static app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes.PREP_SNAPSHOT_ERROR_CODE;
 
-import app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes;
 import app.metatron.discovery.domain.dataprep.exceptions.PrepException;
 import app.metatron.discovery.domain.dataprep.exceptions.PrepMessageKey;
 import java.io.File;
@@ -92,11 +93,47 @@ public class PrepUtil {
     return PrepException.create(PREP_DATASET_ERROR_CODE, msgKey, detail);
   }
 
+  public static PrepException datasetError(PrepMessageKey msgKey) {
+    return PrepException.create(PREP_DATASET_ERROR_CODE, msgKey, null);
+  }
+
   public static PrepException datasetError(Exception e) {
     return PrepException.create(PREP_DATASET_ERROR_CODE, e);
   }
 
+  public static PrepException dataflowError(PrepMessageKey msgKey, String detail) {
+    return PrepException.create(PREP_DATAFLOW_ERROR_CODE, msgKey, detail);
+  }
+
+  public static PrepException dataflowError(PrepMessageKey msgKey) {
+    return PrepException.create(PREP_DATAFLOW_ERROR_CODE, msgKey, null);
+  }
+
   public static PrepException dataflowError(Exception e) {
     return PrepException.create(PREP_DATAFLOW_ERROR_CODE, e);
+  }
+
+  public static PrepException snapshotError(PrepMessageKey msgKey, String detail) {
+    return PrepException.create(PREP_SNAPSHOT_ERROR_CODE, msgKey, detail);
+  }
+
+  public static PrepException snapshotError(PrepMessageKey msgKey) {
+    return PrepException.create(PREP_SNAPSHOT_ERROR_CODE, msgKey, null);
+  }
+
+  public static PrepException snapshotError(Exception e) {
+    return PrepException.create(PREP_SNAPSHOT_ERROR_CODE, e);
+  }
+
+  public static PrepException configError(PrepMessageKey msgKey, String detail) {
+    return PrepException.create(PREP_INVALID_CONFIG_CODE, msgKey, detail);
+  }
+
+  public static PrepException configError(PrepMessageKey msgKey) {
+    return PrepException.create(PREP_INVALID_CONFIG_CODE, msgKey, null);
+  }
+
+  public static PrepException configError(Exception e) {
+    return PrepException.create(PREP_INVALID_CONFIG_CODE, e);
   }
 }
