+++ discovery-server/src/main/java/app/metatron/discovery/domain/dataprep/file/PrepJsonUtil.java
@@ -1,7 +1,10 @@
-package app.metatron.discovery.domain.dataprep.json;
+package app.metatron.discovery.domain.dataprep.file;
 
 import static app.metatron.discovery.domain.dataprep.PrepProperties.HADOOP_CONF_DIR;
+import static app.metatron.discovery.domain.dataprep.PrepUtil.datasetError;
+import static app.metatron.discovery.domain.dataprep.exceptions.PrepMessageKey.MSG_DP_ALERT_FAILED_TO_PARSE_JSON;
 
+import app.metatron.discovery.common.GlobalObjectMapper;
 import app.metatron.discovery.domain.dataprep.exceptions.PrepErrorCodes;
 import app.metatron.discovery.domain.dataprep.exceptions.PrepException;
 import app.metatron.discovery.domain.dataprep.exceptions.PrepMessageKey;
@@ -12,13 +15,12 @@ import com.google.common.collect.Lists;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
+import java.io.Reader;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.sql.ResultSet;
@@ -28,8 +30,6 @@ import java.util.HashMap;
 import java.util.Map;
 import javax.servlet.ServletOutputStream;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.ContentSummary;
-import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
@@ -40,97 +40,20 @@ public class PrepJsonUtil {
 
   private static Logger LOGGER = LoggerFactory.getLogger(PrepJsonUtil.class);
 
-  /**
-   * @param strUri URI as String (to be java.net.URI)
-   * @param limitRows Read not more than this
-   * @param conf Hadoop configuration which is mandatory when the url's protocol is hdfs
-   * @param onlyCount If true, just fill result.totalRows and result.totalBytes
-   * @return PrepCsvParseResult: grid, header, maxColCnt
-   *
-   * Sorry for so many try-catches. Sacrificed readability for end-users' usability.
-   */
-  public static PrepJsonParseResult parseJson(String strUri, int limitRows, Integer columnCount, Configuration conf,
-          boolean onlyCount) {
-    PrepJsonParseResult result = new PrepJsonParseResult();
-    BufferedReader reader;
-    URI uri;
 
-    LOGGER.debug("PrepSnapshotService.parseJson(): strUri={} conf={}", strUri, conf);
+  private static void readJson(Reader reader, int limitRows, Integer manualColCnt, boolean onlyCount,
+          PrepParseResult result) {
+    Integer colCnt = manualColCnt != null ? manualColCnt : null;
 
-    try {
-      uri = new URI(strUri);
-    } catch (URISyntaxException e) {
-      e.printStackTrace();
-      throw PrepException
-              .create(PrepErrorCodes.PREP_DATASET_ERROR_CODE, PrepMessageKey.MSG_DP_ALERT_MALFORMED_URI_SYNTAX, strUri);
-    }
+    ObjectMapper mapper = GlobalObjectMapper.getDefaultMapper();
+    String line;
 
-    switch (uri.getScheme()) {
-      case "hdfs":
-        if (conf == null) {
-          throw PrepException.create(PrepErrorCodes.PREP_INVALID_CONFIG_CODE,
-                  PrepMessageKey.MSG_DP_ALERT_REQUIRED_PROPERTY_MISSING, HADOOP_CONF_DIR);
-        }
-        Path path = new Path(uri);
-
-        FileSystem hdfsFs;
-        try {
-          hdfsFs = FileSystem.get(conf);
-        } catch (IOException e) {
-          e.printStackTrace();
-          throw PrepException.create(PrepErrorCodes.PREP_DATASET_ERROR_CODE,
-                  PrepMessageKey.MSG_DP_ALERT_CANNOT_GET_HDFS_FILE_SYSTEM, strUri);
-        }
-
-        FSDataInputStream his;
-        try {
-          if (onlyCount) {
-            ContentSummary cSummary = hdfsFs.getContentSummary(path);
-            result.totalBytes = cSummary.getLength();
-          }
-
-          his = hdfsFs.open(path);
-        } catch (IOException e) {
-          e.printStackTrace();
-          throw PrepException.create(PrepErrorCodes.PREP_DATASET_ERROR_CODE,
-                  PrepMessageKey.MSG_DP_ALERT_CANNOT_READ_FROM_HDFS_PATH, strUri);
-        }
-
-        reader = new BufferedReader(new InputStreamReader(his));
-        break;
-
-      case "file":
-        File file = new File(uri);
-        if (onlyCount) {
-          result.totalBytes = file.length();
-        }
-
-        FileInputStream fis;
-        try {
-          fis = new FileInputStream(file);
-        } catch (FileNotFoundException e) {
-          e.printStackTrace();
-          throw PrepException.create(PrepErrorCodes.PREP_DATASET_ERROR_CODE,
-                  PrepMessageKey.MSG_DP_ALERT_CANNOT_READ_FROM_LOCAL_PATH, strUri);
-        }
-
-        reader = new BufferedReader(new InputStreamReader(fis));
-        break;
-
-      default:
-        throw PrepException
-                .create(PrepErrorCodes.PREP_DATASET_ERROR_CODE, PrepMessageKey.MSG_DP_ALERT_UNSUPPORTED_URI_SCHEME,
-                        strUri);
-    }
-
-    ObjectMapper mapper = new ObjectMapper();
-    String line = "";
+    BufferedReader br = new BufferedReader(reader);
 
     try {
       result.colNames = Lists.newArrayList();
-      result.maxColCnt = 0;
       StringBuffer sb = new StringBuffer();
-      while ((line = reader.readLine()) != null && result.totalRows < limitRows) {
+      while ((line = br.readLine()) != null && result.totalRows < limitRows) {
         Map<String, Object> jsonRow = null;
 
         try {
@@ -154,14 +77,12 @@ public class PrepJsonUtil {
           }
         }
 
-        int colCnt = result.colNames.size();
-        if (columnCount != null) {
-          colCnt = columnCount;
+        if (colCnt == null) {
+          colCnt = result.colNames.size();
         }
-        result.maxColCnt = colCnt;
 
-        String[] row = new String[result.maxColCnt];
-        for (int i = 0; i < result.maxColCnt; i++) {
+        String[] row = new String[colCnt];
+        for (int i = 0; i < colCnt; i++) {
           String colName = result.colNames.get(i);
           if (jsonRow.containsKey(colName) == true) {
             Object obj = jsonRow.get(colName);
@@ -172,15 +93,33 @@ public class PrepJsonUtil {
       }
     } catch (IOException e) {
       e.printStackTrace();
-      throw PrepException
-              .create(PrepErrorCodes.PREP_DATASET_ERROR_CODE, PrepMessageKey.MSG_DP_ALERT_FAILED_TO_PARSE_JSON, strUri);
+      throw datasetError(MSG_DP_ALERT_FAILED_TO_PARSE_JSON);
     }
+
+  }
+
+  /**
+   * @param strUri URI as String (to be java.net.URI)
+   * @param limitRows Read not more than this
+   * @param manualColCnt Manually set column count from UI
+   * @param conf Hadoop configuration which is mandatory when the url's protocol is hdfs
+   * @param onlyCount If true, just fill result.totalRows and result.totalBytes
+   * @return PrepParseResult: grid, colNames
+   */
+  public static PrepParseResult parse(String strUri, int limitRows, Integer manualColCnt, Configuration conf,
+          boolean onlyCount) {
+    PrepParseResult result = new PrepParseResult();
+
+    LOGGER.debug("PrepJsonUtil.parse(): strUri={} conf={}", strUri, conf);
+
+    Reader reader = PrepFileUtil.getReader(strUri, conf, onlyCount, result);
+    readJson(reader, limitRows, manualColCnt, onlyCount, result);
 
     return result;
   }
 
-  public static PrepJsonParseResult parseJson(String strUri, int limitRows, Integer columnCount, Configuration conf) {
-    return parseJson(strUri, limitRows, columnCount, conf, false);
+  public static PrepParseResult parse(String strUri, int limitRows, Integer manualColCnt, Configuration conf) {
+    return parse(strUri, limitRows, manualColCnt, conf, false);
   }
 
   /**
@@ -193,7 +132,7 @@ public class PrepJsonUtil {
    */
   public static Map<String, Long> countJson(String strUri, int limitRows, Configuration conf) {
     Map<String, Long> mapTotal = new HashMap();
-    PrepJsonParseResult result = parseJson(strUri, limitRows, null, conf, true);
+    PrepParseResult result = parse(strUri, limitRows, null, conf, true);
     mapTotal.put("totalRows", result.totalRows);
     mapTotal.put("totalBytes", result.totalBytes);
     return mapTotal;
