+++ discovery-server/src/main/java/app/metatron/discovery/domain/dataprep/file/PrepParseResult.java
@@ -1,22 +1,19 @@
-package app.metatron.discovery.domain.dataprep.csv;
+package app.metatron.discovery.domain.dataprep.file;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class PrepCsvParseResult {
+public class PrepParseResult {
 
   public List<String[]> grid;
   public List<String> colNames;
-  public int maxColCnt;
 
   // Used only in PrepCsvUtil.parse() when onlyCount is true
   public long totalRows;
   public long totalBytes;
 
-  public PrepCsvParseResult() {
+  public PrepParseResult() {
     grid = new ArrayList();
-    colNames = null;
-    maxColCnt = -1;
     totalRows = 0L;
     totalBytes = 0L;
   }
