+++ discovery-server/src/test/java/app/metatron/discovery/domain/dataprep/csv/ApacheCommonsCsvInputTest.java
@@ -3,6 +3,8 @@ package app.metatron.discovery.domain.dataprep.csv;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import app.metatron.discovery.domain.dataprep.file.PrepParseResult;
+import app.metatron.discovery.domain.dataprep.file.PrepCsvUtil;
 import app.metatron.discovery.domain.dataprep.teddy.DataFrame;
 import org.junit.Test;
 
@@ -17,7 +19,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_with_df() {
     String strUri = buildStrUrlFromResourceDir("csv/sale.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
@@ -26,7 +28,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_bom() {
     String strUri = buildStrUrlFromResourceDir("csv/sale_bom16.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
@@ -35,7 +37,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_bom_and_header() {
     String strUri = buildStrUrlFromResourceDir("csv/sale_bom16.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 30, null, true);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 30, null, true);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
@@ -44,7 +46,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_unstructured() {
     String strUri = buildStrUrlFromResourceDir("csv/unstructured.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show(50);
@@ -53,7 +55,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_multi_line() {
     String strUri = buildStrUrlFromResourceDir("csv/multi_line.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
@@ -64,7 +66,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_multi_line_escape() {
     String strUri = buildStrUrlFromResourceDir("csv/multi_line_with_bs_escape.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
@@ -76,7 +78,7 @@ public class ApacheCommonsCsvInputTest {
   @Test
   public void test_multi_line_unmatched() {
     String strUri = buildStrUrlFromResourceDir("csv/multi_line_with_bs_escape.csv");
-    PrepCsvParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null, null, true, false);
+    PrepParseResult result = PrepCsvUtil.parse(strUri, ",", 10000, null, null, true, false);
     DataFrame df = new DataFrame();
     df.setByGrid(result);
     df.show();
